+++ libs/stack/stack-ui/src/providers/SidePanel/interface.ts
@@ -1,7 +1,9 @@
 import type { OverlayTriggerState } from '@react-stately/overlays'
+import type { DOMProps } from '@react-types/shared'
 import type React from 'react'
+import type { OverlayTriggerProps } from 'react-aria'
 
-export type TSidePanelProviderProps = {
+export type TSidePanelProviderProps = Partial<OverlayTriggerProps> & {
   children: React.ReactNode
   defaultSelectedKey: string
   defaultIsOpen?: boolean
@@ -20,4 +22,5 @@ export type TSidePanelContext = {
   defaultSelectedKey: string
   overlayState: OverlayTriggerState
   buttonProps: TButtonProps
+  overlayProps: DOMProps
 }